Move common code
