+++ src/Common/assertProcessUserMatchesDataOwner.h
@@ -1,0 +1,10 @@
+#pragma once
+#include <string>
+
+namespace DB
+{
+
+void assertProcessUserMatchesDataOwner(
+    const std::string & path, std::function<void(const std::string &)> on_warning);
+
+}
