+++ src/Common/makeSocketAddress.h
@@ -1,0 +1,8 @@
+#include <Poco/Net/SocketAddress.h>
+
+namespace Poco { class Logger; }
+
+namespace DB
+{
+Poco::Net::SocketAddress makeSocketAddress(const std::string & host, uint16_t port, Poco::Logger * log);
+}
